Move byte-allignment and typedarray files to utils
