+++ src/engine/input/RingBuffer.ts
@@ -17,7 +17,7 @@ import {
   writeFloat32,
   writeUint8,
 } from "../allocator/CursorView";
-import { TypedArrayConstructor } from "../allocator/types";
+import { TypedArrayConstructor } from "../utils/typedarray";
 
 export interface InputRingBuffer<T extends TypedArrayConstructor> extends RingBuffer<T> {
   buffer: ArrayBuffer;
