+++ src/engine/utils/byte-alignment.ts
